Create Assignment_Four.pde
This is my fourth Processing file that I made for Intro to MAGD class. I am proud of how I was successfully able to get the large sphere to move by the arrow keys and how I was able to get all the colors to switch at the click of the mouse.